requirements.txt
# Core Dependencies - Required
streamlit>=1.28.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
joblib>=1.3.0

# Optional - TensorFlow (bisa skip jika error, sistem akan pakai sklearn)
# tensorflow>=2.13.0
